refactored delete_book to alert user if book cannot be found

+++ app/books/books_blueprint.py
@@ -137,12 +137,14 @@ def delete_book(uid):
     # have json object from input
     request_json = request.get_json()
 
-    # add changes to session and save them to the db
-    db.session.delete(requested_book)
-    db.session.commit()
+    if request_json:
+        # add changes to session and save them to the db
+        db.session.delete(requested_book)
+        db.session.commit()
 
-    # show updated version of the book
-    return jsonify(requested_book.serialized())
+        # show updated version of the book
+        return jsonify(requested_book.serialized())
+    return "couldn't find that book..."
 
 
 @books_blueprint.route("purge", methods=["DELETE"])
